requirements.txt
Django==5.2.10
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
requests==2.32.3
